tuna: Fix TypeError: 'type' object is not iterable
in getopt.getopt we were passing "int" as the variable long perhaps
as a result of the automatic conversions in 2to3

Fix this

+++ tuna-cmd.py
@@ -468,7 +468,7 @@ def main():
                 if have_inet_diag:
                         short += "n"
                         int.append("show_sockets")
-                opts, args = getopt.getopt(sys.argv[1:], short, int)
+                opts, args = getopt.getopt(sys.argv[1:], short, long)
         except getopt.GetoptError as err:
                 usage()
                 print(str(err))
